# Implement layout="zip" for Lambda/GCF, deprecating lambdex (Cherry-pick of #19076)