… src/python/pants/backend/awslambda/python/rules_test.py
… src/python/pants/backend/python/util_rules/faas_test.py
… src/python/pants/backend/python/util_rules/pex_venv_test.py